requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
et_xmlfile==2.0.0
Flask==3.1.1
Flask-WTF==1.2.2
geographiclib==2.0
geopandas==1.1.1
geopy==2.4.1
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
lxml==6.0.0
MarkupSafe==3.0.2
matching==1.4.3
MouseInfo==0.1.3
narwhals==2.0.1
networkx==3.5
numpy==2.3.2
oauthlib==3.3.1
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
paramiko==3.5.1
pillow==11.3.0
plotly==6.2.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
pygsheets==2.0.6
pylabels==1.2.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyogrio==0.11.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyperclip==1.9.0
pyproj==3.7.1
PyRect==0.2.0
PyScreeze==1.0.1
pysftp==0.2.9
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-pptx==1.0.2
pytweening==1.2.0
pytz==2025.2
qrcode==8.2
reportlab==4.4.3
reportlab-qrcode==0.1.6
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
rubicon-objc==0.5.1
scikit-learn==1.7.1
scipy==1.16.1
shapely==2.1.1
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
xlsxwriter==3.2.5
